loraRecv/requirements.txt
numpy==1.24.4
scipy==1.10.1
pyserial==3.5
pytest==7.0.1
pytest-mock==3.10.0
matplotlib==3.7.1
setuptools>=65.5.1
wheel>=0.38.4